% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{addCells.layers}
\alias{addCells.layers}
\title{Add missing cells to data added to /matrix}
\usage{
addCells.layers(x, n, m2, do.transpose)
}
\arguments{
\item{x}{A list of matricies to add to /layers.}

\item{n}{The number of genes each cell needs.}

\item{m2}{The number of cells being added to the loom file.}

\item{do.transpose}{Logical, whether the matrix data should be transposed before
added to the loom object.}
}
\value{
'x' with the proper number of cells
}
\description{
Add missing cells to data added to /matrix
}
