% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check.layers}
\alias{check.layers}
\title{Check additions to /layers}
\usage{
check.layers(x, n, layers.names, do.transpose)
}
\arguments{
\item{x}{A list of matrices to add layers in /layers.}

\item{n}{The number of genes needed for each layer.}

\item{layers.names}{Names found in /layers.}

\item{do.transpose}{Logical, whether the matrix data should be transposed before
added to the loom object.}
}
\value{
'x' as a list of matricies with 'n' rows for each layer present in /layers
}
\description{
Check additions to /layers
}
