% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{combine}
\alias{combine}
\title{Combine loom files}
\usage{
combine(looms, filename, chunk.size = 1000, order.by = NULL,
  overwrite = FALSE, verbose = TRUE, chunk.dims = NULL, ...)
}
\arguments{
\item{looms}{A list of \code{loom} objects or paths to loom files}

\item{filename}{Name for resultant loom file}

\item{chunk.size}{How many rows from each input loom should we stream to the merged loom file at any given time?}

\item{order.by}{Optional row attribute to order each input loom by, must be one dimensional}

\item{overwrite}{Overwrite \code{filename} if already exists?}

\item{verbose}{Display a progress bar as we're copying over data}

\item{chunk.dims}{Chunk dimensions used for matrix; will be copied from first file by default}

\item{...}{Ignored for now}
}
\value{
A loom object connected to \code{filename}
}
\description{
Combine loom files
}
\seealso{
\code{\link{loom}}
}
