% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{connect}
\alias{connect}
\title{Connect to a loom file}
\usage{
connect(filename, mode = "r+")
}
\arguments{
\item{filename}{The loom file to connect to}

\item{mode}{How do we connect to it? Pass 'r' for read-only or 'r+' for read/write}
}
\value{
A loom file connection
}
\description{
Connect to a loom file
}
