% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create}
\alias{create}
\title{Create a loom object}
\usage{
create(filename, data, row.attrs, col.attrs)
}
\arguments{
\item{filename}{...}

\item{data}{...}

\item{row.attrs}{...}

\item{col.attrs}{...}
}
\value{
A connection to a loom file
}
\description{
Create a loom object
}
\seealso{
\code{\link{loom-class}}
}
