% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\docType{class}
\name{loom-class}
\alias{loom-class}
\alias{loom}
\title{A class for loom}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
loom
}
\value{
Object of class \code{\link{loom}}
}
\description{
A class for loom
}
\seealso{
\code{\link{hdf5r::H5File}}
}
\keyword{datasets}
