% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{map}
\alias{map}
\title{Map a function or a series of functions over a loom file}
\usage{
map(self, f_list = list(mean, var), MARGIN = 1, chunksize = 1000,
  selection)
}
\arguments{
\item{MARGIN}{The dimmension to map over, pass 1 for cells or 2 for genes}

\item{X}{A loom object}

\item{FUN}{A function to map to the loom file}

\item{chunk.size}{Chunk size to use, defaults to \code{loomfile$chunksize[MARGIN]}}

\item{index.use}{Indices of the dataset to use, defaults to \code{1:loomfile$shape[MARGIN]}}

\item{dataset.use}{Dataset to use, defauts to 'matrix'}

\item{display.progress}{Display a progress bar}

\item{expected}{...}

\item{...}{Extra parameters for FUN}
}
\value{
The results of the map
}
\description{
Map a function or a series of functions over a loom file
}
