% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{nCells.layers}
\alias{nCells.layers}
\title{Get the number of cells being added to /layers}
\usage{
nCells.layers(x, do.transpose)
}
\arguments{
\item{x}{A list of matricies to add to /layers.}

\item{do.transpose}{Logical, whether the matrix data should be transposed before
added to the loom object.}
}
\value{
The number of cells within each matrix.
}
\description{
Get the number of cells being added to /layers
}
