% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{nCells.matrix_data}
\alias{nCells.matrix_data}
\title{Get the number of cells being added to /matrix}
\usage{
nCells.matrix_data(x, do.transpose)
}
\arguments{
\item{x}{A matrix to add to /matrix.}

\item{do.transpose}{Logical, whether the matrix data should be transposed before
added to the loom object.}
}
\value{
The number of cells in x.
}
\description{
Get the number of cells being added to /matrix
}
