% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{subset.loom}
\alias{subset.loom}
\title{Subset a loom file}
\usage{
\method{subset}{loom}(x, m, n, filename, overwrite = FALSE,
  display.progress = TRUE, ...)
}
\arguments{
\item{x}{A loom object}

\item{m}{Rows (genes) to subset}

\item{n}{Columns (cells) to subset}

\item{filename}{Filename for new loom object}

\item{overwrite}{Overwrite \code{filename} if already exists?}

\item{display.progress}{Display progress as we're copying over data}
}
\value{
A loom object connected to \code{filename}
}
\description{
Subset a loom file
}
