% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{validateLoom}
\alias{validateLoom}
\title{Validate a loom object}
\usage{
validateLoom(object)
}
\arguments{
\item{object}{A loom object}
}
\value{
None, errors if object is an invalid loom connection
}
\description{
Validate a loom object
}
