% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create}
\alias{create}
\title{Create a loom object}
\usage{
create(filename, data, gene.attrs = NULL, cell.attrs = NULL,
  layers = NULL, chunk.dims = "auto", chunk.size = 1000,
  overwrite = FALSE, display.progress = TRUE)
}
\arguments{
\item{filename}{The name of the new loom file}

\item{data}{The data for \code{/matrix}, should be cells as rows and genes as columns}

\item{gene.attrs}{A named list of vectors with extra data for genes, each vector must be as long as the number of genes in \code{data}}

\item{cell.attrs}{A named list of vectors with extra data for cells, each vector must be as long as the number of cells in \code{data}}

\item{chunk.dims}{A one- or two-length integer vector of chunksizes for \code{/matrix}, defaults to 'auto' to automatically determine chunksize}

\item{chunk.size}{How many rows of \code{data} should we stream to the loom file at any given time?}

\item{overwrite}{Overwrite an already existing loom file?}
}
\value{
A connection to a loom file
}
\description{
Create a loom object
}
\seealso{
\code{\link{loom-class}}
}
