% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{create}
\alias{create}
\title{Create a loom object}
\usage{
create(filename, data, gene.attrs = NULL, cell.attrs = NULL,
  layers = NULL, chunk.dims = "auto", chunk.size = 1000,
  do.transpose = TRUE, calc.numi = FALSE, overwrite = FALSE,
  display.progress = TRUE)
}
\arguments{
\item{filename}{The name of the new loom file}

\item{data}{The data for \code{/matrix}. If cells are rows and genes are columns, set \code{do.transpose = FALSE}; otherwise, set \code{do.transpose = TRUE}}

\item{gene.attrs}{A named list of vectors with extra data for genes, each vector must be as long as the number of genes in \code{data}}

\item{cell.attrs}{A named list of vectors with extra data for cells, each vector must be as long as the number of cells in \code{data}}

\item{layers}{A named list of matrices to be added as layers}

\item{chunk.dims}{A one- or two-length integer vector of chunksizes for \code{/matrix}, defaults to 'auto' to automatically determine chunksize}

\item{chunk.size}{How many rows of \code{data} should we stream to the loom file at any given time?}

\item{do.transpose}{Transpose the input? Should be \code{TRUE} if \code{data} has genes as rows and cells as columns}

\item{calc.numi}{Calculate number of UMIs and genes expressed per cell? Will store in 'col_attrs/nUMI' and 'col_attrs/nGene', overwriting anything passed to \code{cel.attrs};
To set a custom threshold for gene expression, pass an integer value (eg. \code{calc.numi = 5} for a threshold of five counts per cell)}

\item{overwrite}{Overwrite an already existing loom file?}

\item{display.progress}{Display a progress bar}
}
\value{
A connection to a loom file
}
\description{
Create a loom object
}
\seealso{
\code{\link{loom}}
}
