% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\docType{class}
\name{loom-class}
\alias{loom-class}
\alias{loom}
\title{A class for connections loom files}
\format{An \code{\link{R6::R6Class}} object}
\usage{
loom
}
\description{
A class for connections loom files
}
\section{Fields}{

\describe{
\item{\code{version}}{Version of loomR object was created under}

\item{\code{shape}}{Shape of \code{/matrix} in genes (columns) by cells (rows)}

\item{\code{matrix}}{The main data matrix, stored as columns (cells) by rows (genes)}

\item{\code{layers}}{Additional data matricies, the same shape as \code{/matrix}}

\item{\code{col.attrs}}{Extra information about cells}

\item{\code{row.attrs}}{Extra information about genes}
}}

\section{Methods}{

\describe{
  \item{\code{add.graph(a, b, w, name, MARGIN, overwrite)}, \code{add.graph.matrix(mat, name, MARGIN, overwrite)}}{
    Add a graph to the loom object; can add either in coorindate format (\code{add.graph}) or matrix format (\code{add.graph.matrix}).
    Stores graph in coordinate format as \code{[row, col]_graphs/name/a} (row indices),
    \code{[row, col]_graphs/name/b} (column indices), and \code{[row, col]_graphs/name/w} (values)
    \describe{
      \item{\code{a}}{Integer vector of row indices for graph, must be the same lengths as \code{b} and \code{w}}
      \item{\code{b}}{Integer vector of column indices for graph, must be the same lengths as \code{a} and \code{w}}
      \item{\code{w}}{Numeric vector of values for graph, must be the same lengths as \code{a} and \code{b}}
      \item{\code{mat}}{Graph provided as a matrix (sparse or dense) or data.frame}
      \item{\code{name}}{Name to store graph, will end up being \code{col_graphs/name} or \code{row_graphs/name}, depending on \code{MARGIN}}
      \item{\code{MARGIN}}{Store the graph in \code{row_graphs} (1) or \code{col_graphs} (2), defaults to 2}
      \item{\code{overwrite}}{Can overwrite existing graph?}
    }
  }
  \item{\code{add.layer(layer, chunk.size, overwrite)}}{
    Add a data layer to this loom file, must be the same dimensions as \code{/matrix}
    \describe{
      \item{\code{layer}}{A named list of matrices to be added as layers}
      \item{\code{chunk.size}}{Number of rows from each layer to stream at once, defaults to 1000}
      \item{\code{overwrite}}{If a layer already exists, overwrite with new data, defaults to FALSE}
    }
  }
  \item{\code{add.attribute(attribute, MARGIN, overwrite)}}{
    Add extra information to this loom file.
    \describe{
      \item{\code{attribute}}{A named list where the first dimmension of each element as long as one dimension of \code{/matrix}}
      \item{\code{MARGIN}}{Either 1 for genes or 2 for cells}
      \item{\code{overwrite}}{Can overwrite existing attributes?}
    }
  }
  \item{\code{add.row.attribute(attribute), add.col.attribute(attribute)}}{
    Add row or column attributes
  }
  \item{\code{get.attribute.df(MARGIN, attribute.names, row.names, col.names)}}{
    Get a group of row or column attributes as a data frame, will only return attributes that have one dimension
    \describe{
      \item{\code{MARGIN}}{Either '1' or '2' to get row- or column-attributes, respectively}
      \item{\code{attribute.names}}{A vector of attribute dataset basenames}
      \item{\code{row.names}}{Basename of the rownames dataset}
      \item{\code{col.names}}{Basename of the colnames dataset}
    }
  }
  \item{\code{get.graph(name, MARGIN)}}{
    Get a graph as a sparse matrix
    \describe{
      \item{\code{name}}{Name of the graph, can be either the basename or full name of the grpah}
      \item{\code{MARGIN}}{
        Load the graph from \code{row_graphs} (1) or \code{col_graphs} (2), defaults to 2.
        Ignored if full path to graph is passed to \code{name}
      }
    }
  }
  \item{\code{chunk.indices(chunk.size = NULL, MARGIN = 2, dataset.use = 'matrix')}}{
    Generate indices of chunk size \code{chunk.size} for iteration
    \describe{
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self[[dataset.use]]$chunk_dims}}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{dataset.use}}{Name of dataset to use, can be the name, not \code{group/name}, unless the name is present in multiple groups}
    }
  }
  \item{\code{scan(dataset.use = 'matrix', MARGIN = 2, chunk.size = NULL, index.use = NULL, display.progress = TRUE)}}{
    Scan through a dataset
    \describe{
      \item{\code{dataset.use}}{Name of dataset to use, can be the name, not \code{group/name}, unless the name is present in multiple groups}
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self$chunksize}}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{index.use}}{Which specific values of \code{dataset.use} to use, defaults to \code{1:self$shape[MARGIN]} (all values)}
      \item{\code{display.progress}}{Display a progress bar}
    }
  }
  \item{\code{batch.scan(chunk.size, MARGIN, index.use, dataset.use, force.reset)}, \code{batch.next(return.data)}}{
    Scan a dataset in the loom file from \code{index.use[1]} to \code{index.use[2]}, iterating by \code{chunk.size}.
    \describe{
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self$chunksize}}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{index.use}}{Which specific values of \code{dataset.use} to use, defaults to \code{1:self$shape[MARGIN]} (all values)}
      \item{\code{dataset.use}}{Name of dataset to use, can be the name, not \code{group/name}, unless the name is present in multiple groups}
      \item{\code{force.reset}}{Force a reset of the internal iterator}
      \item{\code{return.data}}{Return data for a given chunk, if \code{FALSE}, returns the indices across \code{MARGIN} for said chunk}
    }
  }
  \item{\code{apply(name, FUN, MARGIN, chunk.size, dataset.use, overwrite, display.progress, ...)}}{
    Apply a function over a dataset within the loom file, stores the results in the loom file. Will not make multidimensional attributes.
    \describe{
      \item{\code{name}}{Full name ('group/name') of dataset to store results to}
      \item{\code{FUN}}{Function to apply}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{index.use}}{Which specific values of \code{dataset.use} to use, defaults to \code{1:self$shape[MARGIN]} (all values)}
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self$chunksize}}
      \item{\code{dataset.use}}{Name of dataset to use}
      \item{\code{overwrite}}{Overite \code{name} if already exists}
      \item{\code{display.progress}}{Display progress}
      \item{\code{max.size}}{Maximum HDF5 chunk size}
      \item{\code{chunk.dims}}{HDF5 chunk dimensions}
      \item{\code{display.progress}}{Display a progress bar}
      \item{\code{...}}{Extra parameters to pass to \code{FUN}}
    }
  }
  \item{\code{map(FUN, MARGIN, chunk.size, index.use, dataset.use, display.progress, expected, ...)}}{
    Map a function onto a dataset within the loom file, returns the result into R.
    \describe{
      \item{\code{FUN}}{}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self$chunksize}}
      \item{\code{index.use}}{Which specific values of \code{dataset.use} to use, defaults to \code{1:self$shape[MARGIN]} (all values)}
      \item{\code{dataset.use}}{Name of dataset to use}
      \item{\code{display.progress}}{Display a progress bar}
      \item{\code{...}}{Extra parameters to pass to \code{FUN}}
    }
  }
  \item{\code{add.cells(matrix.data, attributes.data = NULL, layers.data = NULL, display.progress = TRUE)}}{
    Add cells to a loom file.
    \describe{
      \item{\code{matrix.data}}{A list of m2 cells where each entry is a vector of length n (num genes, \code{self$shape[1]})}
      \item{\code{attributes.data}}{A list where each entry is named for one of the datasets in \code{self[['col_attrs']]}; each entry is a vector of length m2.}
      \item{\code{layers.data}}{A list where each entry is named for one of the datasets in \code{self[['layers']]}; each entry is an n-by-m2 matrix where n is the number of genes in this loom file and m2 is the number of cells being added.}
      \item{\code{display.progress}}{Display a progress bar}
    }
  }
  \item{\code{add.loom(other, other.key, self.key, ...)}}{
    Add the contents of another loom file to this one.
    \describe{
      \item{\code{other}}{An object of class \code{loom} or a filename of another loom file}
      \item{\code{other.key}}{Row attribute in \code{other} to add by}
      \item{\code{self.key}}{Row attribute in this loom file to add by}
      \item{\code{...}}{Ignored for now}
    }
  }
  \item{\code{timestamp(x)}}{
    Timestamp a loom object (group or dataset)
    \describe{
      \item{\code{x}}{A character vector of groups or datasets to timestamp}
    }
  }
  \item{\code{last.modified(x = 'self')}}{
    Get the time an object (group, dataset, file) was most recently modified
    \describe{
      \item{\code{x}}{A character vector of groups or datasets to get the timestamp of, defaults to 'self' or the file as a whole; pass \code{NULL} to get all timestamps present}
    }
  }
}
}

\seealso{
\code{\link{loomR}}, \code{\link{hdf5r::H5File}}
}
\keyword{datasets}
