% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\docType{class}
\name{loom-class}
\alias{loom-class}
\alias{loom}
\title{A class for loom files}
\format{An \code{\link{R6::R6Class}} object}
\usage{
loom
}
\value{
Object of \code{\link{R6::R6Class}} to generate \code{loom} objects
}
\description{
A class for loom files
}
\section{Fields}{

\describe{
\item{\code{version}}{Version of loomR object was created under}

\item{\code{shape}}{Shape of \code{/matrix} in genes (columns) by cells (rows)}

\item{\code{chunksize}}{Chunks set for this dataset in columns (cells) by rows (genes)}

\item{\code{matrix}}{The main data matrix, stored as columns (cells) by rows (genes)}

\item{\code{layers}}{Additional data matricies, the same shape as \code{/matrix}}

\item{\code{col.attrs}}{Extra information about cells}

\item{\code{row.attrs}}{Extra information about genes}
}}

\section{Methods}{

\describe{
  \item{\code{add.layer(layer, chunk.size, overwrite)}}{
    Add a data layer to this loom file, must be the same dimensions as \code{/matrix}
    \describe{
      \item{\code{layer}}{A named list of matrices to be added as layers}
      \item{\code{chunk.size}}{Number of rows from each layer to stream at once, defaults to 1000}
      \item{\code{overwrite}}{If a layer already exists, overwrite with new data, defaults to FALSE}
    }
  }
  \item{\code{add.attribute(attribute, MARGIN, overwrite)}}{
    Add extra information to this loom file.
    \describe{
      \item{\code{attribute}}{A named list where the first dimmension of each element as long as one dimension of \code{/matrix}}
      \item{\code{MARGIN}}{Either 1 for genes or 2 for cells}
      \item{\code{overwrite}}{Can overwrite existing attributes?}
    }
  }
  \item{\code{add.row.attribute(attribute), add.col.attribute(attribute)}}{
    Add row or column attributes
  }
  \item{\code{batch.scan(chunk.size, MARGIN, index.use, dataset.use, force.reset)}, \code{batch.next(return.data)}}{
    Scan a dataset in the loom file from \code{index.use[1]} to \code{index.use[2]}, iterating by \code{chunk.size}.
    \describe{
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self$chunksize}}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{index.use}}{Which specific values of \code{dataset.use} to use, defaults to \code{1:self$shape[MARGIN]} (all values)}
      \item{\code{dataset.use}}{Name of dataset to use, can be the name, not \code{group/name}, unless the name is present in multiple groups}
      \item{\code{force.reset}}{Force a reset of the internal iterator}
    }
  }
  \item{\code{apply(name, FUN, MARGIN, chunk.size, dataset.use, overwrite, display.progress, ...)}}{
    Apply a function over a dataset within the loom file, stores the results in the loom file. Will not make multidimensional attributes.
    \describe{
      \item{\code{name}}{Full name ('group/name')of dataset to store results to}
      \item{\code{FUN}}{Function to apply}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{index.use}}{Which specific values of \code{dataset.use} to use, defaults to \code{1:self$shape[MARGIN]} (all values)}
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self$chunksize}}
      \item{\code{dataset.use}}{Name of dataset to use}
      \item{\code{overwrite}}{Overite \code{name} if already exists}
      \item{\code{display.progress}}{Display progress}
      \item{\code{...}}{Extra parameters to pass to \code{FUN}}
    }
  }
  \item{\code{map(FUN, MARGIN, chunk.size, index.use, dataset.use, display.progress, expected, ...)}}{
    Map a function onto a dataset within the loom file, returns the result into R.
    The result will default to the shape of the dataset used; to change pass either 'vector' or 'matrix' to \code{expected}.
    \describe{
      \item{\code{FUN}}{}
      \item{\code{MARGIN}}{Iterate over genes (1) or cells (2), defaults to 2}
      \item{\code{chunk.size}}{Size to chunk \code{MARGIN} by, defaults to \code{self$chunksize}}
      \item{\code{index.use}}{Which specific values of \code{dataset.use} to use, defaults to \code{1:self$shape[MARGIN]} (all values)}
      \item{\code{dataset.use}}{Name of dataset to use}
      \item{\code{display.progress}}{Display progress}
      \item{\code{expected}}{Class of expected result ('matrix' or 'vector'), defaults to class of \code{dataset.use}}
    }
  }
  \item{\code{add.cells(matrix.data, attributes.data = NULL, layers.data = NULL, display.progress = TRUE)}}{
    Add m2 cells to a loom file.
    \describe{
      \item{\code{matrix.data}}{a list of m2 cells where each entry is a vector of length n (num genes, \code{self$shape[1]})}
      \item{\code{attributes.data}}{a list where each entry is named for one of the datasets in \code{self[['col_attrs']]}; each entry is a vector of length m2.}
      \item{\code{layers.data}}{a list where each entry is named for one of the datasets in \code{self[['layers']]}; each entry is an n-by-m2 matrix where n is the number of genes in this loom file and m2 is the number of cells being added.}
      \item{\code{display.progress}}{display progress}
    }
  }
}
}

\seealso{
\code{\link{hdf5r::H5File}}
}
\keyword{datasets}
