% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\docType{class}
\name{loom-class}
\alias{loom-class}
\alias{loom}
\title{A class for loom files}
\format{An \code{\link{R6::R6Class}} object}
\usage{
loom
}
\value{
Object of \code{\link{R6::R6Class}} to generate \code{loom} objects
}
\description{
A class for loom files
}
\section{Fields}{

\describe{
\item{\code{version}}{Version of loomR object was created under}

\item{\code{shape}}{Shape of \code{/matrix} in columns (cells) by rows (genes)}

\item{\code{chunksize}}{Chunks set for this dataset in columns (cells) by rows (genes)}

\item{\code{matrix}}{The main data matrix, stored as columns (cells) by rows (genes)}

\item{\code{layers}}{Additional data matricies, the same shape as \code{/matrix}}

\item{\code{col.attrs}}{Extra information about cells}

\item{\code{row.attrs}}{Extra information about genes}
}}

\section{Methods}{

\describe{
  \item{\code{add.layer(layer)}}{Add a data layer to this loom file, must be in column (cells) by row (genes) orientation}
  \item{\code{add.attribute(attribute, MARGIN)}}{
    Add extra information to this loom file where
    \code{attribute} is a named list where each element is a vector that is as long as one dimension of \code{/matrix} and
    \code{MARGIN} is either 1 for cells or 2 for genes
  }
  \item{\code{add.row.attribute(attribute)}}{A wrapper for \code{add.attribute(attribute, MARGIN = 2)}}
  \item{\code{add.col.attribute(attribute)}}{A wrapper for \code{add.attribute(attribute, MARGIN = 1)}}
  \item{\code{add.meta.data(meta.data)}}{A wrapper for \code{add.attribute(attribute, MARGIN = 1)}}
  \item{\code{batch.scan(chunk.size, MARGIN, index.use, dataset.use, force.reset)}, \code{batch.next()}}{
    Scan a dataset in the loom file from \code{index.use[1]} to \code{index.use[2]}, iterating by \code{chunk.size}.
    \code{dataset.use} can be the name, not \code{group/name}, unless the name is present in multiple groups.
    Pass \code{MARGIN = 1} to iterate on cells or \code{MARGIN = 2} to iterate on genes for 'matrix' or any dataset in 'layers'.
    To force reevaluation of the iterator object, pass \code{force.reset = TRUE}.
    \code{MARGIN} does not need to be set for datasets in 'row_attrs' or 'col_attrs'.
    \code{chunk.size} defaults to \code{self$chunksize}, \code{MARGIN} defaults to 1,
    \code{index.use} defaults to \code{1:self$shape[MARGIN]}, \code{dataset.use} defaults to 'matrix'
  }
  \item{\code{apply(name, FUN, MARGIN, chunk.size, index.use, dataset.use, expected, ...)}}{...}
  \item{\code{map(FUN, MARGIN, chunk.size, index.use, dataset.use, expected, ...)}}{...}
}
}

\seealso{
\code{\link{hdf5r::H5File}}
}
\keyword{datasets}
