% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{subset.loom}
\alias{subset.loom}
\alias{subset}
\title{Subset a loom file}
\usage{
\method{subset}{loom}(x, m = NULL, n = NULL, filename = NULL,
  overwrite = FALSE, display.progress = TRUE, ...)
}
\arguments{
\item{x}{A loom object}

\item{m}{Rows (genes) to subset, defaults to all rows}

\item{n}{Columns (cells) to subset, defaults to all columns}

\item{filename}{Filename for new loom object, defaults to ...}

\item{overwrite}{Overwrite \code{filename} if already exists?}

\item{display.progress}{Display progress as we're copying over data}

\item{...}{Ignored for now}
}
\value{
A loom object connected to \code{filename}
}
\description{
Subset a loom file
}
