% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{subset.loom}
\alias{subset.loom}
\alias{subset}
\title{Subset a loom file}
\usage{
\method{subset}{loom}(x, m = NULL, n = NULL, filename = NULL,
  chunk.size = 1000, overwrite = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{A loom object}

\item{m}{Rows (cells) to subset, defaults to all rows}

\item{n}{Columns (features) to subset, defaults to all columns}

\item{filename}{Filename for new loom object, defaults to ...}

\item{chunk.size}{Chunk size to iterate through \code{x}}

\item{overwrite}{Overwrite \code{filename} if already exists?}

\item{verbose}{Display a progress bar}

\item{...}{Ignored for now}
}
\value{
A loom object connected to \code{filename}
}
\description{
Subset a loom file
}
\seealso{
\code{\link{loom}}
}
